+++ pkg/providers/networklists/testdata/TestDSNetworkList/match_by_id.tf
@@ -1,0 +1,9 @@
+provider "akamai" {
+  edgerc = "~/.edgerc"
+}
+
+
+data "akamai_networklist_network_lists" "test" {
+   name = "40996_ARTYLABWHITELIST"
+   type = "IP"
+}