+++ pkg/providers/networklists/testdata/TestResNetworkListSubscription/match_by_id.tf
@@ -1,0 +1,10 @@
+provider "akamai" {
+  edgerc = "~/.edgerc"
+}
+
+
+resource "akamai_networklist_subscription" "test" {
+   recipients = ["test@email.com"]
+    network_list = ["79536_MARTINNETWORKLIST"] 
+   }
+